fix(wasi-http): validate ports in p2 set_authority like p3
p3 already rejects empty, non-numeric, and out-of-range ports via
parse_authority. p2 only called http::uri::Authority::from_str, which
accepts values such as example.com: and example.com:abc. With those
authorities, default_send_request appends the scheme default port and
builds broken connect targets (example.com::443 / example.com:abc:443).

Share parse_authority for p2 and p3, and extend the unit coverage.
