+++ crates/wasi-http/src/p3/host/types.rs
@@ -116,25 +116,6 @@ fn delete_request_options(
         .context("failed to delete request options from table")
 }
 
-/// Parse an outgoing request `authority`, rejecting a malformed value.
-///
-/// `http::uri::Authority` accepts an authority whose port section is empty or
-/// non-numeric (for example `example.com:` or `example.com:abc`), so the port
-/// is validated here as well. A `:` inside an IPv6 literal host such as `[::1]`
-/// is part of the host rather than a port delimiter, so the port is only looked
-/// for after any closing bracket.
-fn parse_authority(authority: String) -> Result<http::uri::Authority, ()> {
-    let has_port = match authority.rfind(']') {
-        Some(i) => authority[i..].contains(':'),
-        None => authority.contains(':'),
-    };
-    let authority = http::uri::Authority::try_from(authority).map_err(|_| ())?;
-    if has_port && authority.port_u16().is_none() {
-        return Err(());
-    }
-    Ok(authority)
-}
-
 impl HostFields for WasiHttpCtxView<'_> {
     fn new(&mut self) -> wasmtime::Result<Resource<Fields>> {
         push_fields(self.table, FieldMap::new_mutable(self.ctx.field_size_limit))
@@ -366,7 +347,7 @@ impl HostRequest for WasiHttpCtxView<'_> {
             req.authority = None;
             return Ok(Ok(()));
         };
-        let Ok(authority) = parse_authority(authority) else {
+        let Ok(authority) = crate::parse_authority(authority) else {
             return Ok(Err(()));
         };
         req.authority = Some(authority);
@@ -585,25 +566,3 @@ impl Host for WasiHttpCtxView<'_> {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn authority_accepts_ipv6_and_validates_ports() {
-        use super::parse_authority;
-
-        // Host names and IPv4 literals, with and without an explicit port.
-        assert!(parse_authority("example.com".into()).is_ok());
-        assert!(parse_authority("example.com:443".into()).is_ok());
-        assert!(parse_authority("127.0.0.1:80".into()).is_ok());
-
-        // Bracketed IPv6 literals: the colons belong to the host, so a missing
-        // port must still be accepted and not mistaken for an empty port.
-        assert!(parse_authority("[::1]".into()).is_ok());
-        assert!(parse_authority("[2001:db8::1]".into()).is_ok());
-        assert!(parse_authority("[::1]:443".into()).is_ok());
-
-        // When a port section is present it must be a valid number.
-        assert!(parse_authority("example.com:".into()).is_err());
-        assert!(parse_authority("example.com:abc".into()).is_err());
-    }
-}
